@@ -1,51 +1,90 @@
-"""Command-line interface for the Estate PDF Organizer."""
+"""Command line interface for the Estate PDF Organizer."""
 
-import click
+import argparse
+import os
 from pathlib import Path
-from typing import Optional
-from loguru import logger
 
-@click.group()
-def cli():
-    """Estate PDF Organizer - Process and organize estate documents using LLMs."""
-    pass
+from estate_pdf_organizer.classifier import LLMClassifier
 
-@cli.command()
-@click.argument('input-dir', type=click.Path(exists=True, file_okay=False, dir_okay=True, path_type=Path))
-@click.argument('taxonomy', type=click.Path(exists=True, file_okay=True, dir_okay=False, path_type=Path))
-@click.option('--output-dir', '-o', type=click.Path(file_okay=False, dir_okay=True, path_type=Path),
-              help='Directory to store organized documents')
-@click.option('--dry-run', is_flag=True, help='Preview changes without making them')
-@click.option('--overwrite', is_flag=True, help='Overwrite existing files in output directory')
-@click.option('--log-level', type=click.Choice(['DEBUG', 'INFO', 'WARNING', 'ERROR']),
-              default='INFO', help='Logging level')
-def process(input_dir: Path, taxonomy: Path, output_dir: Optional[Path], 
-           dry_run: bool, overwrite: bool, log_level: str):
-    """Process and organize PDF documents from INPUT_DIR using the taxonomy in TAXONOMY."""
-    # Configure logging
-    logger.remove()
-    logger.add(lambda msg: click.echo(msg), level=log_level)
+from .processor import EstatePDFProcessor
+
+
+def main():
+    """Main entry point for the CLI."""
+    parser = argparse.ArgumentParser(
+        description="Organize estate documents by classifying them."
+    )
     
-    try:
-        # Set default output directory if not provided
-        if output_dir is None:
-            output_dir = input_dir / "organized"
-            logger.info(f"Using default output directory: {output_dir}")
-            
-        # Process documents
-        logger.info(f"Processing documents from {input_dir}")
-        logger.info(f"Using taxonomy from {taxonomy}")
-        if dry_run:
-            logger.info("Running in dry-run mode")
-        if overwrite:
-            logger.warning("Overwrite mode enabled - existing files will be overwritten")
-            
-        # TODO: Implement actual processing logic
-        logger.info("Processing complete")
-        
-    except Exception as e:
-        logger.error(f"Error processing documents: {str(e)}")
-        raise click.ClickException(str(e))
+    parser.add_argument(
+        "input_dir",
+        type=Path,
+        help="Directory containing input PDFs"
+    )
+    
+    parser.add_argument(
+        "output_dir",
+        type=Path,
+        help="Directory to store organized documents"
+    )
+    
+    parser.add_argument(
+        "--taxonomy",
+        type=Path,
+        required=True,
+        help="Path to YAML file containing document taxonomy"
+    )
+    
+    parser.add_argument(
+        "--openai-api-key",
+        type=str,
+        help="OpenAI API key. If not provided, will use OPENAI_API_KEY environment variable."
+    )
+    
+    parser.add_argument(
+        "--overwrite",
+        action="store_true",
+        help="Overwrite existing files in output directory"
+    )
+    
+    parser.add_argument(
+        "--dry-run",
+        action="store_true",
+        help="Show what would be done without making changes"
+    )
+    
+    parser.add_argument(
+        "--window-size",
+        type=int,
+        default=5,
+        help="Number of pages to consider for document boundary detection (default: 5)"
+    )
+    
+    args = parser.parse_args()
+    
+    # Get OpenAI API key from argument or environment variable
+    api_key = args.openai_api_key or os.environ.get("OPENAI_API_KEY")
+    if not api_key:
+        raise ValueError(
+            "OpenAI API key must be provided either via --openai-api-key argument "
+            "or OPENAI_API_KEY environment variable"
+        )
+    
+    # Create output directory if it doesn't exist
+    args.output_dir.mkdir(parents=True, exist_ok=True)
+    
+    # Initialize the processor
+    classifier = LLMClassifier(args.taxonomy, api_key=api_key)
+    processor = EstatePDFProcessor(
+        input_dir=args.input_dir,
+        output_dir=args.output_dir,
+        classifier=classifier,
+        overwrite=args.overwrite,
+        dry_run=args.dry_run,
+        window_size=args.window_size
+    )
+    
+    # Process the PDFs
+    processor.process_directory()
 
-if __name__ == '__main__':
-    cli() 
+if __name__ == "__main__":
+    main() 
