@@ -1,37 +1,43 @@
-"""Tests for the organizer module."""
+"""Tests for the document organizer module."""
 
-from pathlib import Path
 import tempfile
+from pathlib import Path
+
 import yaml
-from pypdf import PdfWriter, PdfReader
-import pdfplumber
-import pytest
-from estate_pdf_organizer.organizer import DocumentOrganizer, DocumentMetadata
+from pypdf import PdfReader, PdfWriter
+
+from estate_pdf_organizer.organizer import DocumentMetadata, DocumentOrganizer
+
 
 def test_document_metadata():
     """Test DocumentMetadata dataclass."""
     metadata = DocumentMetadata(
-        original_pdf="test.pdf",
+        source_pdf="test.pdf",
         start_page=1,
         end_page=3,
-        category="Important Documents",
+        document_type="Important Documents",
         filename="will.pdf",
         confidence=0.95
     )
-    assert metadata.original_pdf == "test.pdf"
+    
+    assert metadata.source_pdf == "test.pdf"
     assert metadata.start_page == 1
     assert metadata.end_page == 3
-    assert metadata.category == "Important Documents"
+    assert metadata.document_type == "Important Documents"
     assert metadata.filename == "will.pdf"
     assert metadata.confidence == 0.95
     assert metadata.output_path is None
 
 def test_document_organizer_initialization():
-    """Test DocumentOrganizer initialization."""
+    """Test initializing the document organizer."""
     with tempfile.TemporaryDirectory() as temp_dir:
         organizer = DocumentOrganizer(Path(temp_dir))
         assert organizer.output_dir == Path(temp_dir)
+        assert not organizer.overwrite
         assert len(organizer.metadata) == 0
+        
+        # Verify output directory was created
+        assert Path(temp_dir).exists()
         assert (Path(temp_dir) / "Unorganized").exists()
 
 def test_add_document():
@@ -39,13 +45,14 @@ def test_add_document():
     with tempfile.TemporaryDirectory() as temp_dir:
         organizer = DocumentOrganizer(Path(temp_dir))
         metadata = DocumentMetadata(
-            original_pdf="test.pdf",
+            source_pdf="test.pdf",
             start_page=1,
             end_page=3,
-            category="Important Documents",
+            document_type="Important Documents",
             filename="will.pdf",
             confidence=0.95
         )
+        
         organizer.add_document(metadata)
         assert len(organizer.metadata) == 1
         assert organizer.metadata[0] == metadata
@@ -60,155 +67,216 @@ def test_organize_document_dry_run():
             writer.add_blank_page(width=612, height=792)  # Standard letter size
         with open(pdf_path, 'wb') as f:
             writer.write(f)
-    
-        organizer = DocumentOrganizer(Path(output_dir))
-        page_ranges = [(1, 3)]
-        document_type = "Important Documents"
         
-        results = organizer.organize_document(
-            source_pdf=str(pdf_path),
-            page_ranges=page_ranges,
-            document_type=document_type,
-            dry_run=True
-        )
-        
-        # Verify metadata was created but no PDF files were written
-        assert len(results) == 1
-        assert results[0].original_pdf == str(pdf_path)
-        assert results[0].start_page == 1
-        assert results[0].end_page == 3
-        assert results[0].category == document_type
-        assert len(list(Path(output_dir).glob("**/*.pdf"))) == 0  # No PDF files should be created
+        # Open the PDF for reading
+        with open(pdf_path, 'rb') as f:
+            pdf_reader = PdfReader(f)
+            
+            organizer = DocumentOrganizer(Path(output_dir))
+            document_type = "Important Documents"
+            
+            result = organizer.organize_document(
+                pdf_reader=pdf_reader,
+                source_pdf_path=str(pdf_path),
+                start_page=1,
+                end_page=3,
+                document_type=document_type,
+                dry_run=True
+            )
+            
+            # Verify metadata was created but no PDF files were written
+            assert result.source_pdf == str(pdf_path)
+            assert result.start_page == 1
+            assert result.end_page == 3
+            assert result.document_type == document_type
+            assert not any(Path(output_dir).glob("**/*.pdf"))
 
 def test_organize_document_with_real_pdf():
-    """Test document organization with a real PDF file."""
+    """Test organizing a document with a real PDF file."""
     with tempfile.TemporaryDirectory() as input_dir, tempfile.TemporaryDirectory() as output_dir:
         # Create a simple PDF file with 3 pages
         pdf_path = Path(input_dir) / "test.pdf"
         writer = PdfWriter()
         for i in range(3):
-            # Add a blank page
             writer.add_blank_page(width=612, height=792)  # Standard letter size
         with open(pdf_path, 'wb') as f:
             writer.write(f)
-    
-        organizer = DocumentOrganizer(Path(output_dir))
-        page_ranges = [(1, 2)]  # Extract first two pages
-        document_type = "Important Documents"
         
-        # Extract pages
-        results = organizer.organize_document(
-            source_pdf=str(pdf_path),
-            page_ranges=page_ranges,
-            document_type=document_type
-        )
-        
-        # Verify the output
-        assert len(results) == 1
-        output_file = Path(results[0].output_path)
-        assert output_file.exists()
-        
-        # Verify the extracted PDF has 2 pages
-        reader = PdfReader(str(output_file))
-        assert len(reader.pages) == 2
+        # Open the PDF for reading
+        with open(pdf_path, 'rb') as f:
+            pdf_reader = PdfReader(f)
+            
+            organizer = DocumentOrganizer(Path(output_dir))
+            document_type = "Important Documents"
+            
+            result = organizer.organize_document(
+                pdf_reader=pdf_reader,
+                source_pdf_path=str(pdf_path),
+                start_page=1,
+                end_page=3,
+                document_type=document_type
+            )
+            
+            # Verify metadata was created
+            assert result.source_pdf == str(pdf_path)
+            assert result.start_page == 1
+            assert result.end_page == 3
+            assert result.document_type == document_type
+            
+            # Verify PDF file was created
+            output_path = Path(result.output_path)
+            assert output_path.exists()
+            
+            # Verify PDF content
+            with open(output_path, 'rb') as f:
+                output_reader = PdfReader(f)
+                assert len(output_reader.pages) == 3
 
 def test_save_metadata():
     """Test saving metadata to YAML file."""
     with tempfile.TemporaryDirectory() as temp_dir:
         organizer = DocumentOrganizer(Path(temp_dir))
         metadata = DocumentMetadata(
-            original_pdf="test.pdf",
+            source_pdf="test.pdf",
             start_page=1,
             end_page=3,
-            category="Important Documents",
+            document_type="Important Documents",
             filename="will.pdf",
             confidence=0.95,
             output_path="output/Important Documents/will.pdf"
         )
-        organizer.add_document(metadata)
-        organizer.save_metadata()
         
+        organizer.add_document(metadata)
+        
+        # Save metadata
         metadata_path = Path(temp_dir) / "metadata.yaml"
+        organizer.save_metadata(metadata_path)
+        
+        # Verify metadata file was created
         assert metadata_path.exists()
         
+        # Load and verify metadata
         with open(metadata_path) as f:
-            saved_metadata = yaml.safe_load(f)
-            assert len(saved_metadata["documents"]) == 1
-            assert saved_metadata["documents"][0]["original_pdf"] == "test.pdf"
-            assert saved_metadata["documents"][0]["start_page"] == 1
-            assert saved_metadata["documents"][0]["end_page"] == 3
-            assert saved_metadata["documents"][0]["category"] == "Important Documents"
-            assert saved_metadata["documents"][0]["filename"] == "will.pdf"
-            assert saved_metadata["documents"][0]["confidence"] == 0.95
-            assert saved_metadata["documents"][0]["output_path"] == "output/Important Documents/will.pdf"
+            loaded_metadata = yaml.safe_load(f)
+        
+        assert "documents" in loaded_metadata
+        assert len(loaded_metadata["documents"]) == 1
+        doc = loaded_metadata["documents"][0]
+        assert doc["source_pdf"] == metadata.source_pdf
+        assert doc["start_page"] == metadata.start_page
+        assert doc["end_page"] == metadata.end_page
+        assert doc["document_type"] == metadata.document_type
+        assert doc["filename"] == metadata.filename
+        assert doc["confidence"] == metadata.confidence
+        assert doc["output_path"] == metadata.output_path
 
 def test_organize_document_invalid_page_range():
-    """Test document organization with invalid page ranges."""
+    """Test organizing a document with invalid page ranges."""
     with tempfile.TemporaryDirectory() as input_dir, tempfile.TemporaryDirectory() as output_dir:
         # Create a simple PDF file with 3 pages
         pdf_path = Path(input_dir) / "test.pdf"
         writer = PdfWriter()
         for i in range(3):
-            writer.add_blank_page(width=612, height=792)
+            writer.add_blank_page(width=612, height=792)  # Standard letter size
         with open(pdf_path, 'wb') as f:
             writer.write(f)
-    
-        organizer = DocumentOrganizer(Path(output_dir))
         
-        # Test start page < 1
-        with pytest.raises(ValueError, match="Start page must be at least 1"):
-            organizer.organize_document(
-                source_pdf=str(pdf_path),
-                page_ranges=[(0, 2)],  # Invalid start page
-                document_type="Important Documents"
-            )
-        
-        # Test end page > total pages
-        with pytest.raises(ValueError, match="End page 4 exceeds total pages"):
-            organizer.organize_document(
-                source_pdf=str(pdf_path),
-                page_ranges=[(1, 4)],  # Invalid end page
-                document_type="Important Documents"
-            )
-        
-        # Test start page > end page
-        with pytest.raises(ValueError, match="Start page 3 is greater than end page 2"):
-            organizer.organize_document(
-                source_pdf=str(pdf_path),
-                page_ranges=[(3, 2)],  # Invalid range
-                document_type="Important Documents"
-            )
+        # Open the PDF for reading
+        with open(pdf_path, 'rb') as f:
+            pdf_reader = PdfReader(f)
+            
+            organizer = DocumentOrganizer(Path(output_dir))
+            document_type = "Important Documents"
+            
+            # Test start page < 1
+            try:
+                organizer.organize_document(
+                    pdf_reader=pdf_reader,
+                    source_pdf_path=str(pdf_path),
+                    start_page=0,
+                    end_page=3,
+                    document_type=document_type
+                )
+                assert False, "Expected ValueError for start_page < 1"
+            except ValueError:
+                pass
+            
+            # Test end page > total pages
+            try:
+                organizer.organize_document(
+                    pdf_reader=pdf_reader,
+                    source_pdf_path=str(pdf_path),
+                    start_page=1,
+                    end_page=4,
+                    document_type=document_type
+                )
+                assert False, "Expected ValueError for end_page > total pages"
+            except ValueError:
+                pass
+            
+            # Test start page > end page
+            try:
+                organizer.organize_document(
+                    pdf_reader=pdf_reader,
+                    source_pdf_path=str(pdf_path),
+                    start_page=3,
+                    end_page=1,
+                    document_type=document_type
+                )
+                assert False, "Expected ValueError for start_page > end_page"
+            except ValueError:
+                pass
 
 def test_organize_document_with_existing_output():
-    """Test document organization when output file already exists."""
+    """Test organizing a document when output file already exists."""
     with tempfile.TemporaryDirectory() as input_dir, tempfile.TemporaryDirectory() as output_dir:
-        # Create a simple PDF file
+        # Create a simple PDF file with 3 pages
         pdf_path = Path(input_dir) / "test.pdf"
         writer = PdfWriter()
-        writer.add_blank_page(width=612, height=792)
+        for i in range(3):
+            writer.add_blank_page(width=612, height=792)  # Standard letter size
         with open(pdf_path, 'wb') as f:
             writer.write(f)
-    
-        # Create output file
-        output_path = Path(output_dir) / "test_pages_1-1.pdf"
-        output_path.touch()
-    
-        organizer = DocumentOrganizer(Path(output_dir), overwrite=False)
         
-        # Should raise error when overwrite=False
-        with pytest.raises(ValueError, match="Output file already exists"):
-            organizer.organize_document(
-                source_pdf=str(pdf_path),
-                page_ranges=[(1, 1)],
-                document_type="Important Documents"
+        # Create output directory and file
+        document_type = "Important Documents"
+        output_dir_path = Path(output_dir) / document_type
+        output_dir_path.mkdir(parents=True)
+        output_path = output_dir_path / "test_pages_1-3.pdf"
+        with open(output_path, 'w') as f:
+            f.write("dummy file")
+        
+        # Open the PDF for reading
+        with open(pdf_path, 'rb') as f:
+            pdf_reader = PdfReader(f)
+            
+            organizer = DocumentOrganizer(Path(output_dir))
+            
+            # Test without overwrite
+            try:
+                organizer.organize_document(
+                    pdf_reader=pdf_reader,
+                    source_pdf_path=str(pdf_path),
+                    start_page=1,
+                    end_page=3,
+                    document_type=document_type
+                )
+                assert False, "Expected ValueError for existing output file"
+            except ValueError:
+                pass
+            
+            # Test with overwrite
+            organizer = DocumentOrganizer(Path(output_dir), overwrite=True)
+            result = organizer.organize_document(
+                pdf_reader=pdf_reader,
+                source_pdf_path=str(pdf_path),
+                start_page=1,
+                end_page=3,
+                document_type=document_type
             )
-        
-        # Should succeed when overwrite=True
-        organizer = DocumentOrganizer(Path(output_dir), overwrite=True)
-        results = organizer.organize_document(
-            source_pdf=str(pdf_path),
-            page_ranges=[(1, 1)],
-            document_type="Important Documents"
-        )
-        assert len(results) == 1
-        assert Path(results[0].output_path).exists() 
+            
+            # Verify file was overwritten
+            assert Path(result.output_path).exists()
+            with open(result.output_path, 'rb') as f:
+                output_reader = PdfReader(f)
+                assert len(output_reader.pages) == 3 
